refactor: removed static name of import github routes
