+++ src/services/github-services/framework-detector-services.js
@@ -1,6 +1,5 @@
 import { AxiosHttpClientAdapter } from '../axios/AxiosHttpClientAdapter'
 import makeGithubApi from '@/services/axios/makeGithubApi'
-import * as Errors from '@/services/axios/errors'
 import { Buffer } from 'buffer'
 import { getVulcanPresets } from '@/helpers'
 
@@ -23,19 +22,10 @@ export const frameworkDetectorService = async ({ accountName, repositoryName }) 
   )
 
   if (httpResponse.body?.content) {
-    httpResponse = adapt(httpResponse)
+    return detectFrameworkByDependencies(httpResponse)
   }
 
-  return parseHttpResponse(httpResponse)
-}
-
-const adapt = (httpResponse) => {
-  const framework = detectFrameworkByDependencies(httpResponse)
-
-  return {
-    body: framework,
-    statusCode: httpResponse.statusCode
-  }
+  return null
 }
 
 const convertPackageToJsonFormat = (httpResponse) => {
@@ -55,29 +45,4 @@ const detectFrameworkByDependencies = (httpResponse) => {
   const hasMatchCases = allDetectedFrameworks.length > 0
 
   return hasMatchCases ? allDetectedFrameworks[0] : null
-}
-
-/**
- * @param {Object} httpResponse - The HTTP response object.
- * @param {Object} httpResponse.body - The response body.
- * @param {String} httpResponse.statusCode - The HTTP status code.
- * @returns {string} The result message based on the status code.
- * @throws {Error} If there is an error with the response.
- */
-const parseHttpResponse = (httpResponse) => {
-  switch (httpResponse.statusCode) {
-    case 200:
-      return httpResponse.body
-    case 401:
-      throw new Errors.InvalidApiTokenError().message
-    case 403:
-      throw new Errors.PermissionError().message
-    case 404:
-      const apiError = 'This repository has no package.json, please try another one.'
-      throw new Error(apiError).message
-    case 500:
-      throw new Errors.InternalServerError().message
-    default:
-      throw new Errors.UnexpectedError().message
-  }
 }
